Update README.md
Hello there! 👋 I'm Jason, and I'd like to welcome you to my GitHub space where data tells the story, and every analysis opens a door to new insights. 

As an aspiring Data Analyst, my journey is filled with curiosity, learning, and a relentless drive to make sense of the ever-growing sea of data.